# internal/server/openapi/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Corvo is a distributed background job queue. All endpoints are under `/api/v1`. Authentication via `X-API-Key` header or `Authorization: Bearer` token.",
        "title": "Corvo API",
        "contact": {
            "url": "https://corvohq.com"
        },
        "version": "v1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/ack/batch": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Worker"
                ],
                "summary": "Acknowledge multiple jobs",
                "parameters": [
                    {
                        "description": "Ack batch request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.AckBatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.AckedCountResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ack/{job_id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Worker"
                ],
                "summary": "Acknowledge (complete) a job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Ack body",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/server.AckBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/enqueue": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Enqueue a job",
                "parameters": [
                    {
                        "description": "Enqueue request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/store.EnqueueRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Duplicate exists (unique_key)",
                        "schema": {
                            "$ref": "#/definitions/store.EnqueueResult"
                        }
                    },
                    "201": {
                        "description": "Job enqueued",
                        "schema": {
                            "$ref": "#/definitions/store.EnqueueResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limited or budget exceeded",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/enqueue/batch": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Enqueue a batch of jobs",
                "parameters": [
                    {
                        "description": "Batch enqueue request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/store.BatchEnqueueRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/store.BatchEnqueueResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/fail/{job_id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Marks the job as failed. If retries remain it will be retried according to the backoff policy.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Worker"
                ],
                "summary": "Fail a job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fail body",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/server.FailBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/fetch": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Long-polls until a job is available or timeout expires. Returns one job or 204 if none available.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Worker"
                ],
                "summary": "Fetch a job (long-poll)",
                "parameters": [
                    {
                        "description": "Fetch request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/store.FetchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.FetchResult"
                        }
                    },
                    "204": {
                        "description": "No job available"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/fetch/batch": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Long-polls until jobs are available or timeout expires. Returns up to count jobs.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Worker"
                ],
                "summary": "Fetch multiple jobs (long-poll)",
                "parameters": [
                    {
                        "description": "Fetch batch request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.FetchBatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.FetchBatchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/heartbeat": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Extends leases for active jobs. Can include progress and checkpoint data. Returns cancellation signals.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Worker"
                ],
                "summary": "Worker heartbeat",
                "parameters": [
                    {
                        "description": "Heartbeat request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/store.HeartbeatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.HeartbeatResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/webhooks/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Delete a webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Webhook ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "server.AckBatchRequest": {
            "type": "object",
            "properties": {
                "acks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.AckBatchItem"
                    }
                }
            }
        },
        "server.AckBody": {
            "type": "object",
            "properties": {
                "agent_status": {
                    "type": "string"
                },
                "checkpoint": {
                    "type": "string"
                },
                "exit_reason": {
                    "type": "string"
                },
                "hold_reason": {
                    "type": "string"
                },
                "result": {
                    "type": "string"
                },
                "step_status": {
                    "type": "string"
                },
                "usage": {
                    "$ref": "#/definitions/store.UsageReport"
                }
            }
        },
        "server.AckedCountResponse": {
            "type": "object",
            "properties": {
                "acked": {
                    "type": "integer"
                }
            }
        },
...
        "server.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "server.StatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
...
        "store.EnqueueRequest": {
            "type": "object",
            "properties": {
                "agent": {
                    "$ref": "#/definitions/store.AgentConfig"
                },
                "chain": {
                    "$ref": "#/definitions/store.ChainDefinition"
                },
                "chain_config": {
                    "type": "string"
                },
                "chain_id": {
                    "type": "string"
                },
                "chain_step": {
                    "type": "integer"
                },
                "checkpoint": {
                    "type": "string"
                },
                "depends_on": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "expire_after": {
                    "description": "e.g. \"1h\"",
                    "type": "string"
                },
                "max_retries": {
                    "type": "integer"
                },
                "parent_id": {
                    "type": "string"
                },
                "payload": {
                    "type": "string"
                },
                "priority": {
                    "type": "string"
                },
                "queue": {
                    "type": "string"
                },
                "retry_backoff": {
                    "type": "string"
                },
                "retry_base_delay": {
                    "description": "e.g. \"5s\"",
                    "type": "string"
                },
                "retry_max_delay": {
                    "description": "e.g. \"10m\"",
                    "type": "string"
                },
                "scheduled_at": {
                    "type": "string"
                },
                "tags": {
                    "type": "string"
                },
                "unique_key": {
                    "type": "string"
                },
                "unique_period": {
                    "description": "seconds",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Bearer token: `Authorization: Bearer \u003ctoken\u003e`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}